Create kingdavid
I want to create a safety software app for all transportation safety. I believe that if technology fail that  they can switch over to fuel engine. or it can have the capacity of hover craft, different capabilities up under the wings the plain, and have  some kind of sensor to alert them of  the problems that the transportation board  will recognize. and to switch from technology to fuel  capacity.